LayoutTests/platform/mac/http/tests/inspector/network/resource-sizes-memory-cache-expected.txt

Test for Resource size values (transfer size, decoded size, header size, request and response) for resources served from the memory cache.


== Running test suite: Resource.Size.MemoryCache
-- Running test case: Resource.Size.MemoryCache.Basic
PASS: statusCode should be 304.
PASS: compressed should be false.
PASS: responseSource should be Symbol(memory-cache).
PASS: estimatedNetworkEncodedSize should be exactly 0 bytes.
PASS: estimatedTotalTransferSize should be >= (encoded body size + headers).
PASS: size should be exactly 27 bytes.
PASS: networkEncodedSize should be exactly 0 bytes.
PASS: networkTotalTransferSize should be >= (encoded body size + headers).
PASS: requestBodyTransferSize should be exactly 0 bytes.
PASS: responseBodyTransferSize should be exactly 0 bytes.
PASS: requestHeadersTransferSize should be non-empty.
PASS: responseHeadersTransferSize should be non-empty.
